test cases that are less broken

+++ test.py
@@ -1,23 +1,40 @@
 import logging
 import subprocess
+import sys
 import time
+import gevent
+from gevent import monkey
+monkey.patch_all()
+
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
 
-from luminoso_api import LuminosoClient, LuminosoAuthError
+from luminoso_api import LuminosoClient
 
 ROOT_CLIENT = None
 PROJECT = None
 USERNAME = None
 
-PROJECT_NAME = 'test5'
+PROJECT_NAME = 'test11'
 ROOT_URL = 'http://localhost:5000/v3'
+
+def fileno_monkeypatch(self):
+    return sys.__stdout__.fileno()
+
+import StringIO
+StringIO.StringIO.fileno = fileno_monkeypatch
 
 def error(obj):
     return obj.get('error')
 
+def run_api_server():
+    from lumi_api.flask_init import init_flask
+    gevent.spawn(init_flask, api_name='localhost:5000')
+    gevent.sleep(2)
+
 def setup():
     global ROOT_CLIENT, PROJECT, USERNAME
+    run_api_server()
     user_info_str = subprocess.check_output('tellme lumi-test', shell=True)
     user_info = eval(user_info_str)
     USERNAME = user_info['username']
@@ -56,14 +73,41 @@ def test_upload():
          'title': 'example-1'},
         {'text': 'Examples are a great source of inspiration',
          'title': 'example-2'},
+        {'text': 'Great things come in threes',
+         'title': 'example-3'},
     ]
-    assert not error(PROJECT.upload('docs', docs))
-    assert not error(PROJECT.post('docs/calculate'))
+    job_id = PROJECT.upload('docs', docs)['result']
+    job_id_2 = PROJECT.post('docs/calculate')
+    assert job_id_2 > job_id
     assert not error(PROJECT.get('docs/calculate'))
     PROJECT.wait_for_assoc()
     assert not error(PROJECT.get('terms'))
 
+def test_topics():
+    topics = PROJECT.get('topics')
+    assert topics['result'] == []
+
+    PROJECT.post('topics',
+        name='Example topic',
+        role='topic',
+        color='#aabbcc',
+        surface_texts=['Examples']
+    )
+
+    topics = PROJECT.get('topics')
+    result = topics['result']
+    assert len(result) == 1
+    topic = result[0]
+    assert topic['name'] == 'Example topic'
+    assert topic['surface_texts'] == ['Examples']
+    assert topic['color'] == '#aabbcc'
+    topic_id = topic['id']
+
+    topic2 = PROJECT.get('topics/id/%s' % topic_id)
+    assert topic2 == topic
+
 def teardown():
-    ROOT_CLIENT.delete(USERNAME + '/projects/' + PROJECT_NAME)
-    PROJECT = ROOT_CLIENT.change_path(USERNAME + '/projects/' + PROJECT_NAME)
-    assert error(PROJECT.get())
+    if ROOT_CLIENT is not None:
+        ROOT_CLIENT.delete(USERNAME + '/projects/' + PROJECT_NAME)
+        PROJECT = ROOT_CLIENT.change_path(USERNAME + '/projects/' + PROJECT_NAME)
+        assert error(PROJECT.get())
